[REFACTOR] HTTP Req/Res 로깅 필터, ClockConfig 리패키징
