+++ src/main/java/kr/hhplus/be/server/config/ClockConfig.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.config.web;
+package kr.hhplus.be.server.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
